move tx ErrorCodes type to core
